Ajuste tabela, models, context e infra

+++ Domain/Model/Quadrant.cs
@@ -2,7 +2,7 @@
 {
     public class Quadrant
     {
-        public short QuadrantId { get; set; }
+        public byte QuadrantId { get; set; }
         public string QuadrantResource { get; set; }
         public virtual ICollection<Place> Places { get; set; }
     }
